Extract cvr validation to seperate module
